+++ scripts/packageJSONLint.ts
@@ -12,6 +12,7 @@ import {
   isMain,
   readFile,
 } from "isaacscript-common-node";
+import type { ReadonlyRecord } from "isaacscript-common-ts";
 import { isKebabCase } from "isaacscript-common-ts";
 import path from "node:path";
 
@@ -60,7 +61,7 @@ export function packageJSONLint(): boolean {
 
 function isPackageJSONValid(
   packageJSONPath: string,
-  rootDeps: Record<string, string> | undefined,
+  rootDeps: ReadonlyRecord<string, string> | undefined,
 ): boolean {
   const isRoot = rootDeps === undefined;
   const isTemplateFile = packageJSONPath.includes("dynamic");
@@ -310,7 +311,7 @@ function getVersionForSpecificPackage(packageName: string): string {
 
 function checkDeps(
   object: unknown,
-  rootDeps: Record<string, string>,
+  rootDeps: ReadonlyRecord<string, string>,
   packageJSONPath: string,
 ): boolean {
   const deps = object as Record<string, string> | undefined | null;
@@ -354,8 +355,8 @@ function checkDeps(
 }
 
 function checkRootDepsUpToDate(
-  rootDeps: Record<string, string>,
-  packageJSONPaths: string[],
+  rootDeps: ReadonlyRecord<string, string>,
+  packageJSONPaths: readonly string[],
 ) {
   for (const [rootDepName, rootDepVersion] of Object.entries(rootDeps)) {
     const matchingPackageJSONPath = packageJSONPaths.find((packageJSONPath) =>